requirements.txt
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
boto3==1.28.62
botocore==1.31.62
click==8.1.7
cloudpickle==2.2.1
contextlib2==21.6.0
dill==0.3.7
exceptiongroup==1.1.3
fastapi==0.101.1
google-pasta==0.2.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
multiprocess==0.70.15
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pathos==0.3.1
Pillow==10.0.0
platformdirs==3.10.0
pox==0.3.3
ppft==1.7.6.7
protobuf==4.24.0
pydantic==2.1.1
pydantic_core==2.4.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
s3transfer==0.7.0
sagemaker==2.177.1
schema==0.7.5
six==1.16.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.0
starlette==0.27.0
tblib==1.7.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
websockets==11.0.3
zipp==3.16.2
